FDG-9822 report filter logic
